QUALITY: Refactoring
